#39 Refactoring of components.
-See title.
-Also added links in current login page to reach homepage without having to manually entering the url. Click on 'Click here' at the bottom :)
-Fixed typos pointed out by @13alan (thanks again!)
