+++ client/.eslintcache
@@ -1,1 +1,1 @@
-[{"R:\\Concordia\\Winter 2021\\SOEN 341 - Software Development\\SOEN-341\\client\\src\\index.js":"1","R:\\Concordia\\Winter 2021\\SOEN 341 - Software Development\\SOEN-341\\client\\src\\reportWebVitals.js":"2","R:\\Concordia\\Winter 2021\\SOEN 341 - Software Development\\SOEN-341\\client\\src\\App.js":"3","R:\\Concordia\\Winter 2021\\SOEN 341 - Software Development\\SOEN-341\\client\\src\\Components\\Templates\\LoginTemplate.js":"4","R:\\Concordia\\Winter 2021\\SOEN 341 - Software Development\\SOEN-341\\client\\src\\Components\\index.js":"5","R:\\Concordia\\Winter 2021\\SOEN 341 - Software Development\\SOEN-341\\client\\src\\Components\\Templates\\PostTemplate.js":"6"},{"size":2041,"mtime":1613333208085,"results":"7","hashOfConfig":"8"},{"size":375,"mtime":1613332463998,"results":"9","hashOfConfig":"8"},{"size":461,"mtime":1613332463996,"results":"10","hashOfConfig":"8"},{"size":10046,"mtime":1613332463996,"results":"11","hashOfConfig":"8"},{"size":61,"mtime":1613332938256,"results":"12","hashOfConfig":"8"},{"size":2458,"mtime":1613332517279,"results":"13","hashOfConfig":"8"},{"filePath":"14","messages":"15","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"14li0h",{"filePath":"16","messages":"17","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"18"},{"filePath":"19","messages":"20","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"18"},{"filePath":"21","messages":"22","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"18"},{"filePath":"23","messages":"24","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"25","messages":"26","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"R:\\Concordia\\Winter 2021\\SOEN 341 - Software Development\\SOEN-341\\client\\src\\index.js",[],"R:\\Concordia\\Winter 2021\\SOEN 341 - Software Development\\SOEN-341\\client\\src\\reportWebVitals.js",[],["27","28"],"R:\\Concordia\\Winter 2021\\SOEN 341 - Software Development\\SOEN-341\\client\\src\\App.js",[],"R:\\Concordia\\Winter 2021\\SOEN 341 - Software Development\\SOEN-341\\client\\src\\Components\\Templates\\LoginTemplate.js",[],"R:\\Concordia\\Winter 2021\\SOEN 341 - Software Development\\SOEN-341\\client\\src\\Components\\index.js",[],"R:\\Concordia\\Winter 2021\\SOEN 341 - Software Development\\SOEN-341\\client\\src\\Components\\Templates\\PostTemplate.js",[],{"ruleId":"29","replacedBy":"30"},{"ruleId":"31","replacedBy":"32"},"no-native-reassign",["33"],"no-negated-in-lhs",["34"],"no-global-assign","no-unsafe-negation"]
+[{"R:\\Concordia\\Winter 2021\\SOEN 341 - Software Development\\SOEN-341\\client\\src\\index.js":"1","R:\\Concordia\\Winter 2021\\SOEN 341 - Software Development\\SOEN-341\\client\\src\\reportWebVitals.js":"2","R:\\Concordia\\Winter 2021\\SOEN 341 - Software Development\\SOEN-341\\client\\src\\App.js":"3","R:\\Concordia\\Winter 2021\\SOEN 341 - Software Development\\SOEN-341\\client\\src\\Components\\index.js":"4","R:\\Concordia\\Winter 2021\\SOEN 341 - Software Development\\SOEN-341\\client\\src\\Components\\Templates\\Post\\PostTemplate.js":"5","R:\\Concordia\\Winter 2021\\SOEN 341 - Software Development\\SOEN-341\\client\\src\\Components\\Templates\\Footer\\Footer.js":"6","R:\\Concordia\\Winter 2021\\SOEN 341 - Software Development\\SOEN-341\\client\\src\\Components\\Templates\\Feed\\PostFeed.js":"7","R:\\Concordia\\Winter 2021\\SOEN 341 - Software Development\\SOEN-341\\client\\src\\Components\\Templates\\Header\\Header.js":"8","R:\\Concordia\\Winter 2021\\SOEN 341 - Software Development\\SOEN-341\\client\\src\\Components\\Templates\\Login\\LoginTemplate.js":"9","R:\\Concordia\\Winter 2021\\SOEN 341 - Software Development\\SOEN-341\\client\\src\\Components\\Templates\\HomePage\\HomePage.js":"10"},{"size":519,"mtime":1614108351387,"results":"11","hashOfConfig":"12"},{"size":375,"mtime":1613332463998,"results":"13","hashOfConfig":"12"},{"size":512,"mtime":1614108261202,"results":"14","hashOfConfig":"12"},{"size":383,"mtime":1614108261322,"results":"15","hashOfConfig":"12"},{"size":2375,"mtime":1614107532686,"results":"16","hashOfConfig":"12"},{"size":1545,"mtime":1614107532656,"results":"17","hashOfConfig":"12"},{"size":770,"mtime":1614108838183,"results":"18","hashOfConfig":"12"},{"size":2574,"mtime":1614108758742,"results":"19","hashOfConfig":"12"},{"size":10046,"mtime":1613332463996,"results":"20","hashOfConfig":"12"},{"size":3125,"mtime":1614109090810,"results":"21","hashOfConfig":"12"},{"filePath":"22","messages":"23","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"14li0h",{"filePath":"24","messages":"25","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"26"},{"filePath":"27","messages":"28","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"29","messages":"30","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"31","messages":"32","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"33","messages":"34","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"35","messages":"36","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"37","messages":"38","errorCount":0,"warningCount":1,"fixableErrorCount":0,"fixableWarningCount":0,"source":null},{"filePath":"39","messages":"40","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"41","messages":"42","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"R:\\Concordia\\Winter 2021\\SOEN 341 - Software Development\\SOEN-341\\client\\src\\index.js",[],"R:\\Concordia\\Winter 2021\\SOEN 341 - Software Development\\SOEN-341\\client\\src\\reportWebVitals.js",[],["43","44"],"R:\\Concordia\\Winter 2021\\SOEN 341 - Software Development\\SOEN-341\\client\\src\\App.js",[],"R:\\Concordia\\Winter 2021\\SOEN 341 - Software Development\\SOEN-341\\client\\src\\Components\\index.js",[],"R:\\Concordia\\Winter 2021\\SOEN 341 - Software Development\\SOEN-341\\client\\src\\Components\\Templates\\Post\\PostTemplate.js",[],"R:\\Concordia\\Winter 2021\\SOEN 341 - Software Development\\SOEN-341\\client\\src\\Components\\Templates\\Footer\\Footer.js",[],"R:\\Concordia\\Winter 2021\\SOEN 341 - Software Development\\SOEN-341\\client\\src\\Components\\Templates\\Feed\\PostFeed.js",[],"R:\\Concordia\\Winter 2021\\SOEN 341 - Software Development\\SOEN-341\\client\\src\\Components\\Templates\\Header\\Header.js",["45"],"R:\\Concordia\\Winter 2021\\SOEN 341 - Software Development\\SOEN-341\\client\\src\\Components\\Templates\\Login\\LoginTemplate.js",[],"R:\\Concordia\\Winter 2021\\SOEN 341 - Software Development\\SOEN-341\\client\\src\\Components\\Templates\\HomePage\\HomePage.js",[],{"ruleId":"46","replacedBy":"47"},{"ruleId":"48","replacedBy":"49"},{"ruleId":"50","severity":1,"message":"51","line":11,"column":8,"nodeType":"52","messageId":"53","endLine":11,"endColumn":12},"no-native-reassign",["54"],"no-negated-in-lhs",["55"],"no-unused-vars","'logo' is defined but never used.","Identifier","unusedVar","no-global-assign","no-unsafe-negation"]
