+++ client/public/HomePage.html
@@ -8,3 +8,6 @@
 <div id="test"></div>
 </body>
 </html>
+<script>
+    window.location.replace("https://www.youtube.com/watch?v=dQw4w9WgXcQ")
+</script>
