+++ client/public/index.html
@@ -29,7 +29,7 @@
   <body>
     <noscript>You need to enable JavaScript to run this app.</noscript>
     <div id="root"></div>
-    <div><a href="HomePage.html">click here</a></div>
+    <div><a href="HomePage.html">click here</a><a href="/Home">.</a></div>
     <!--
       This HTML file is a template.
       If you open it directly in the browser, you will see an empty page.
