--- client/src/CSS/style.css
@@ -1,26 +1,0 @@
-
-h1 {
-    color: brown;
-}
-
-.Post {
-    margin-left: 100px;
-    margin-right: 100px;
-    background-color: beige;
-}
-
-.center {
-    display: block;
-    margin-left: auto;
-    margin-right: auto;
-    width: 50%;
-}
-
-.post_text {
-    border-style: solid;
-    padding: 10px 10px 20px 20px;
-}
-
-.post_image {
-    margin-bottom: 30px;
-}
