+++ client/src/Components/Templates/HomePage/HomePage.js
@@ -5,15 +5,12 @@ import { makeStyles } from '@material-ui/core/styles';
 import CssBaseline from '@material-ui/core/CssBaseline';
 import Grid from '@material-ui/core/Grid';
 import Container from '@material-ui/core/Container';
-import Header from './Header';
-import Main from './Main';
-import Footer from './Footer';
-
-import Post from './PostTemplate';
+import {Post, Header, PostFeed, Footer} from '../../../Components/index';
 //Hard coded test images
 //TODO link image fetch to db
-import pepega from '../../Images/Pepega.png';
-import pikachu from '../../Images/shocked pikachu.png';
+import pepega from '../../../Images/Pepega.png';
+import pikachu from '../../../Images/shocked pikachu.png';
+import videotron from '../../../Images/videotron.png'
 
 const useStyles = makeStyles((theme) => ({
     mainGrid: {
@@ -38,20 +35,29 @@ const sections = [
 //Combine this with backend DB
 const post1 = <Post
     author="Olivier"
+    date="2021-02-12"
+    title="Videotron"
+    imagePath={videotron}
+    text="Haha internet problems on quiz time goes brrrrr"
+/>
+
+const post2 = <Post
+    author="Olivier"
     date="2021-02-11"
     title="Pepega"
     imagePath={pepega}
     text="Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum."
 />
 
-const post2 = <Post
+const post3 = <Post
     author="Olivier"
     date="2021-02-11"
     title="Surprised Pikachu"
     imagePath={pikachu}
     text="Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum."
 />
-const posts = [post1, post2];
+
+const posts = [post1, post2, post3];
 
 
 
@@ -63,13 +69,13 @@ export default function HomePage() {
             <CssBaseline />
             <Container maxWidth="lg">
                 <Header title="MemeSpace" sections={sections} />
-                <main>
-                    <Grid container className={classes.mainGrid}>
-                        <Main title="The Meme Feed" posts={posts} />
+                <feed>
+                    <Grid container className={classes.mainGrid} direction="column">
+                        <PostFeed title="The Meme Feed" posts={posts} />
                     </Grid>
-                </main>
+                </feed>
             </Container>
-            <Footer title="Footer" description="This is a footer" />
+            <Footer title="Footer" description="This is a footer :^)" />
         </React.Fragment>
     );
 }
