--- client/src/Components/Templates/post-test.md
@@ -1,1 +1,0 @@
-test :0
