+++ client/src/Components/index.js
@@ -1,3 +1,8 @@
-import Post from "./Templates/PostTemplate"
+import Post from "./Templates/Post/PostTemplate"
+import PostFeed from "./Templates/Feed/PostFeed";
+import Footer from "./Templates/Footer/Footer";
+import Header from "./Templates/Header/Header";
+import HomePage from "./Templates/HomePage/HomePage";
+import LoginTemplate from "./Templates/Login/LoginTemplate";
 
-export {Post};
+export {Post, PostFeed, Footer, HomePage, Header, LoginTemplate};
